requirements.txt
# ====================================================
# RAG Agent v2 - Production Dependencies
# ====================================================
# Strict version pinning for faster dependency resolution
# Versions match currently installed and tested packages

# Core RAG Framework (LlamaIndex)
llama-index==0.14.10
llama-index-core==0.14.10
llama-index-embeddings-openai==0.5.1
llama-index-llms-openai==0.6.11
llama-index-vector-stores-faiss==0.5.1

# Agent Framework (LangChain) - Strict pinning
langchain==1.1.3
langchain-core==1.2.0
langchain-openai==1.1.3

# Vector Store
faiss-cpu==1.13.1

# OpenAI
openai==2.11.0

# PDF Processing
pypdf==6.4.1

# Environment Management
python-dotenv==1.2.1

# Utilities
numpy==2.3.5
pandas==2.2.3

# Testing & Debugging
jupyter==1.0.0
notebook==7.1.0
ipykernel==7.1.0

# Type Checking & Code Quality
typing-extensions==4.15.0

# Evaluation Layer (LLM-as-a-Judge)
google-generativeai==0.8.3

# GUI App (Streamlit)
streamlit>=1.30.0




